add past paper programs, did minor bug fixes and changed file names to camelCase

+++ inches.c++
@@ -1,17 +1,18 @@
 #include <iostream>
 using namespace std;
 
-// TODO: Write a program that takes feet and calculates and displays the inches and centimeters using two functions 
+// * Write a program that takes feet
+// * calculates and displays the inches and centimeters using two functions 
 
-int toInches(int feet);
-float toCentimeters(int feet);
+float toInches(float feet);
+float toCentimeters(float feet);
 
 int main(){
-    int feet;
+    float feet;
     cout<<"Enter feets: ";
     cin>>feet;
 
-    int inches = toInches(feet);
+    float inches = toInches(feet);
     float centimeters = toCentimeters(feet);
 
     cout<<feet<<" Feet = " << inches << " Inches.\n";
@@ -20,10 +21,10 @@ int main(){
     return 0;
 }
 
-int toInches(int feet){
+float toInches(float feet){
     return feet * 12;
 }
 
-float toCentimeters(int feet){
+float toCentimeters(float feet){
     return feet * 30.48;
 }
